Áñadidas algúnas mejoras
