+++ tests/Feature/Livewire/Admin/ThematicArea/UpdateThematicAreaTest.php
@@ -1,103 +1,14 @@
 <?php
 
-namespace Tests\Feature\Livewire\Admin;
+namespace Tests\Feature\Livewire\Admin\ThematicArea;
 
-use App\Http\Livewire\Admin\ThematicAreas;
+use App\Http\Livewire\Admin\ThematicArea\ThematicAreas;
 use App\Models\ThematicArea;
 use Livewire\Livewire;
 use Tests\TestCase;
 
-class ThematicAreaTest extends TestCase
+class UpdateThematicAreaTest extends TestCase
 {
-    /** @test */
-    public function itShowsTheThematicAreas()
-    {
-        $adminUser = $this->createAdmin();
-        $place = $this->createPlace();
-        $pointOfInterest = $this->createPointOfInterest($place->id);
-
-        $thematicArea1 = $this->createThematicArea($pointOfInterest->id);
-        $thematicArea2 = $this->createThematicArea($pointOfInterest->id);
-
-        $this->actingAs($adminUser);
-
-        $this->get('areas-tematicas')
-            ->assertOk()
-            ->assertSeeInOrder([
-                $thematicArea1->description,
-                $thematicArea2->description,
-            ]);
-    }
-
-    /** @test */
-    public function itCreatesAThematicArea()
-    {
-        $adminUser = $this->createAdmin();
-
-        $this->actingAs($adminUser);
-
-        $this->assertDatabaseCount('thematic_areas', 0);
-
-        Livewire::test(ThematicAreas::class)
-            ->set('createForm.name', 'Nombre')
-            ->set('createForm.description', 'Descripción')
-            ->call('save');
-
-        $this->assertDatabaseCount('thematic_areas', 1);
-
-        $this->assertDatabaseHas('thematic_areas', [
-            'id' => 1,
-            'name' => 'Nombre',
-            'description' => 'Descripción',
-        ]);
-    }
-
-    /** @test */
-    public function itCreatesAThematicAreaWithEmptyName()
-    {
-        $adminUser = $this->createAdmin();
-
-        $this->actingAs($adminUser);
-
-        $this->assertDatabaseCount('thematic_areas', 0);
-
-        Livewire::test(ThematicAreas::class)
-            ->set('createForm.name', '')
-            ->set('createForm.description', 'Descripción')
-            ->call('save');
-
-        $this->assertDatabaseCount('thematic_areas', 1);
-
-        $this->assertDatabaseHas('thematic_areas', [
-            'id' => 1,
-            'name' => '',
-            'description' => 'Descripción',
-        ]);
-    }
-
-    /** @test */
-    public function itCreatesAThematicAreaWithEmptyDescription()
-    {
-        $adminUser = $this->createAdmin();
-
-        $this->actingAs($adminUser);
-
-        $this->assertDatabaseCount('thematic_areas', 0);
-
-        Livewire::test(ThematicAreas::class)
-            ->set('createForm.name', 'Nombre')
-            ->set('createForm.description', '')
-            ->call('save');
-
-        $this->assertDatabaseCount('thematic_areas', 1);
-
-        $this->assertDatabaseHas('thematic_areas', [
-            'id' => 1,
-            'name' => 'Nombre',
-            'description' => '',
-        ]);
-    }
-
     /** @test */
     public function itUpdatesAThematicArea()
     {
@@ -201,15 +112,14 @@ class ThematicAreaTest extends TestCase
     }
 
     /** @test */
-    public function itDeletesAThematicArea()
+    public function itDoesNotUpdateAThematicAreaThatExceedsTheMaxLengthInTheName()
     {
         $adminUser = $this->createAdmin();
-        $place = $this->createPlace();
-        $pointOfInterest = $this->createPointOfInterest($place->id);
 
-        $thematicArea = $this->createThematicArea($pointOfInterest->id);
-
-        $this->actingAs($adminUser);
+        $thematicArea = factory(ThematicArea::class)->create([
+            'name' => 'Nombre',
+            'description' => 'Descripción',
+        ]);
 
         $this->assertDatabaseCount('thematic_areas', 1);
 
@@ -219,9 +129,55 @@ class ThematicAreaTest extends TestCase
             'description' => $thematicArea->description,
         ]);
 
-        Livewire::test(ThematicAreas::class)
-            ->call('delete', $thematicArea);
+        $this->actingAs($adminUser);
 
-        $this->assertDatabaseCount('videos', 0);
+        Livewire::test(ThematicAreas::class)
+            ->set('editForm.name', \str_repeat('a', 46))
+            ->set('editForm.description', 'Descripción actualizada')
+            ->call('update', $thematicArea)
+            ->assertHasErrors(['editForm.name' => 'max']);
+
+        $this->assertDatabaseCount('thematic_areas', 1);
+
+        $this->assertDatabaseHas('thematic_areas', [
+            'id' => $thematicArea->id,
+            'name' => $thematicArea->name,
+            'description' => $thematicArea->description,
+        ]);
+    }
+
+    /** @test */
+    public function itDoesNotUpdateAThematicAreaThatExceedsTheMaxLengthInTheDescription()
+    {
+        $adminUser = $this->createAdmin();
+
+        $thematicArea = factory(ThematicArea::class)->create([
+            'name' => 'Nombre',
+            'description' => 'Descripción',
+        ]);
+
+        $this->assertDatabaseCount('thematic_areas', 1);
+
+        $this->assertDatabaseHas('thematic_areas', [
+            'id' => $thematicArea->id,
+            'name' => $thematicArea->name,
+            'description' => $thematicArea->description,
+        ]);
+
+        $this->actingAs($adminUser);
+
+        Livewire::test(ThematicAreas::class)
+            ->set('editForm.name', 'Nombre actualizado')
+            ->set('editForm.description', \str_repeat('a', 2001))
+            ->call('update', $thematicArea)
+            ->assertHasErrors(['editForm.description' => 'max']);
+
+        $this->assertDatabaseCount('thematic_areas', 1);
+
+        $this->assertDatabaseHas('thematic_areas', [
+            'id' => $thematicArea->id,
+            'name' => $thematicArea->name,
+            'description' => $thematicArea->description,
+        ]);
     }
 }
